Fix context not updating in indicator
