+++ packages/experimental/TabList/src/TabList/TabList.types.ts
@@ -6,19 +6,23 @@ import type { FocusZoneProps } from '@fluentui-react-native/focus-zone';
 import type { LayoutTokens } from '@fluentui-react-native/tokens';
 import type { LayoutRectangle } from '@office-iss/react-native-win32';
 
-import type {
-  AnimatedIndicatorStyles,
-  AnimatedIndicatorStylesUpdate,
-  TabLayoutInfo,
-  TabListAnimatedIndicatorProps,
-} from '../TabListAnimatedIndicator/TabListAnimatedIndicator.types';
+import type { AnimatedIndicatorStyles, AnimatedIndicatorStylesUpdate } from '../TabListAnimatedIndicator/TabListAnimatedIndicator.types';
 
 export const tabListName = 'TabList';
 
 export type TabListAppearance = 'transparent' | 'subtle';
 export type TabListSize = 'small' | 'medium' | 'large';
 
-export interface TabListContextData {
+export interface TabLayoutInfo extends LayoutRectangle {
+  startMargin?: number;
+  tabBorderWidth?: number;
+}
+export interface TabListLayoutInfo {
+  tablist: LayoutRectangle;
+  tabs: { [key: string]: TabLayoutInfo };
+}
+
+export interface TabListState {
   /**
    * Method to add tabKey to context's `tabKeys` list. Run once on a tab mounting.
    */
@@ -52,6 +56,8 @@ export interface TabListContextData {
    * @platform win32
    */
   invoked?: boolean;
+
+  layout?: TabListLayoutInfo;
 
   /**
    * Updates the selected Tab and calls the client’s onTabSelect callback
@@ -88,11 +94,6 @@ export interface TabListContextData {
    * Array of Tab values in the group
    */
   tabKeys: string[];
-
-  /**
-   * Layout information of the TabList, used to help animate the animated indicator.
-   */
-  tabListLayout?: LayoutRectangle;
 
   /**
    * Directly update the animated indicator's styles with styles the user supplies for each slot.
@@ -161,10 +162,6 @@ export interface TabListProps extends Pick<FocusZoneProps, 'isCircularNavigation
    */
   vertical?: boolean;
 }
-
-export interface TabListState {
-  context?: TabListContextData;
-}
 export interface TabListInfo {
   props: TabListProps;
   state: TabListState;
@@ -172,7 +169,6 @@ export interface TabListInfo {
 export interface TabListSlotProps {
   container?: FocusZoneProps;
   stack: React.PropsWithRef<IViewProps>;
-  animatedIndicator?: TabListAnimatedIndicatorProps;
 }
 
 export interface TabListType {
