+++ packages/experimental/TabList/src/TabListAnimatedIndicator/TabListAnimatedIndicator.win32.tsx
@@ -4,10 +4,10 @@ import React from 'react';
 import { View } from 'react-native';
 import type { ViewProps, ViewStyle } from 'react-native';
 
-import { stagedComponent, mergeProps, memoize } from '@fluentui-react-native/framework';
+import { stagedComponent, memoize } from '@fluentui-react-native/framework';
 
-import type { TabListAnimatedIndicatorProps } from './TabListAnimatedIndicator.types';
 import { tablistAnimatedIndicatorName } from './TabListAnimatedIndicator.types';
+import { useAnimatedIndicatorStyles } from './useAnimatedIndicatorStyles';
 
 const getIndicatorProps = memoize(indicatorPropsWorker);
 function indicatorPropsWorker(animationClass: string, style: ViewStyle): ViewProps {
@@ -18,13 +18,12 @@ function indicatorPropsWorker(animationClass: string, style: ViewStyle): ViewPro
  * This component renders as the indicator for the selected tab. Its styles are calculated and passed
  * from the useTabList hook, so it doesn't need to use the compose or compressible franework.
  */
-export const TabListAnimatedIndicator = stagedComponent((props: TabListAnimatedIndicatorProps) => {
-  return (final: TabListAnimatedIndicatorProps) => {
-    const { styles, ...finalProps } = final;
-    const rootProps = mergeProps(props, finalProps, { style: styles.container });
+export const TabListAnimatedIndicator = stagedComponent(() => {
+  const styles = useAnimatedIndicatorStyles();
+  return () => {
     const indicatorProps = getIndicatorProps('Ribbon_TabUnderline', styles.indicator);
     return (
-      <View {...rootProps}>
+      <View style={styles.container}>
         <View {...indicatorProps} />
       </View>
     );
